pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "datasets[vision]>=3.5.0",
    "fastapi[standard]>=0.115.12",
    "ftfy>=6.3.1",
    "regex>=2024.11.6",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "weaviate-client>=4.13.2",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
